Move config loading into the Config class. Self contain dbFacile too. Update rrdtool default setting to be similar to others. Move silly cisco-entities.php into entity-physical.inc.php
