+++ LibreNMS/Config.php
@@ -28,6 +28,64 @@ namespace LibreNMS;
 class Config
 {
     /**
+     * Load the user config from config.php
+     *
+     * @param string $install_dir
+     * @return array
+     */
+    public static function &load($install_dir = null)
+    {
+        global $config;
+
+        if (empty($install_dir)) {
+            $install_dir = __DIR__ . '/../';
+        }
+        $install_dir = realpath($install_dir);
+        $config['install_dir'] = $install_dir;
+
+        // load defaults
+        require $install_dir . '/includes/defaults.inc.php';
+        require $install_dir . '/includes/definitions.inc.php';
+
+        // variable definitions (remove me)
+        require $install_dir . '/includes/vmware_guestid.inc.php';
+
+        // Load user config
+        include $install_dir . '/config.php';
+
+        return $config;
+    }
+
+    /**
+     * Load Config from the database
+     *
+     * @throws Exceptions\DatabaseConnectException
+     */
+    public static function &loadFromDatabase()
+    {
+        global $config;
+
+        if (empty($config)) {
+            self::load();
+        }
+
+        // Make sure the database is connected
+        dbConnect();
+
+        // pull in the database config settings
+        self::mergeDb();
+
+        // load graph types from the database
+        self::loadGraphsFromDb();
+
+        // Process $config to tidy up
+        self::processConfig();
+
+        return $config;
+    }
+
+
+    /**
      * Get a config value, if non existent null (or default if set) will be returned
      *
      * @param string $key period separated config variable name
@@ -223,5 +281,102 @@ class Config
         }
 
         return is_array($curr) && isset($curr[$last]);
+    }
+
+    /**
+     * merge the database config with the global config
+     * Global config overrides db
+     */
+    private static function mergeDb()
+    {
+        global $config;
+
+        $db_config = array();
+        foreach (dbFetchRows('SELECT `config_name`,`config_value` FROM `config`') as $obj) {
+            self::assignArrayByPath($db_config, $obj['config_name'], $obj['config_value']);
+        }
+        $config = array_replace_recursive($db_config, $config);
+    }
+
+    /**
+     * Assign a value into the passed array by a path
+     * 'snmp.version' = 'v1' becomes $arr['snmp']['version'] = 'v1'
+     *
+     * @param array $arr the array to insert the value into, will be modified in place
+     * @param string $path the path to insert the value at
+     * @param mixed $value the value to insert, will be type cast
+     * @param string $separator path separator
+     */
+    private static function assignArrayByPath(&$arr, $path, $value, $separator = '.')
+    {
+        // type cast value. Is this needed here?
+        if (filter_var($value, FILTER_VALIDATE_INT)) {
+            $value = (int) $value;
+        } elseif (filter_var($value, FILTER_VALIDATE_FLOAT)) {
+            $value = (float) $value;
+        } elseif (filter_var($value, FILTER_VALIDATE_BOOLEAN, FILTER_NULL_ON_FAILURE) !== null) {
+            $value = filter_var($value, FILTER_VALIDATE_BOOLEAN);
+        }
+
+        $keys = explode($separator, $path);
+
+        // walk the array creating keys if they don't exist
+        foreach ($keys as $key) {
+            $arr = &$arr[$key];
+        }
+        // assign the variable
+        $arr = $value;
+    }
+
+    private static function loadGraphsFromDb()
+    {
+        global $config;
+
+        // load graph types from the database
+        foreach (dbFetchRows('SELECT * FROM graph_types') as $graph) {
+            foreach ($graph as $k => $v) {
+                if (strpos($k, 'graph_') == 0) {
+                    // remove leading 'graph_' from column name
+                    $key = str_replace('graph_', '', $k);
+                } else {
+                    $key = $k;
+                }
+                $g[$key] = $v;
+            }
+
+            $config['graph_types'][$g['type']][$g['subtype']] = $g;
+        }
+    }
+
+    /**
+     * Proces the config after it has been loaded.
+     * Make sure certain variables have been set properly and
+     *
+     *
+     */
+    private static function processConfig()
+    {
+        if (!self::get('email_from')) {
+            self::set('email_from', '"' . self::get('project_name') . '" <' . self::get('email_user') . '@' . php_uname('n') . '>');
+        }
+
+        if (self::get('secure_cookies')) {
+            ini_set('session.cookie_secure', 1);
+        }
+
+        if ($rrd_real_95th = self::get('rrdgraph_real_95th')) {
+            self::set('rrdgraph_real_percentile', $rrd_real_95th);
+        }
+
+        if ($fping_millisc = self::get('fping_options.millisec')) {
+            self::set('fping_options.interval', $fping_millisc);
+        }
+
+        // make sure we have full path to binaries in case PATH isn't set
+        foreach (array('fping', 'fping6', 'snmpgetnext', 'rrdtool') as $bin) {
+            if (!is_executable(self::get($bin))) {
+                self::set($bin, locate_binary($bin), true, $bin, "Path to $bin", 'external', 'paths');
+            }
+        }
     }
 }
