--- includes/cisco-entities.php
@@ -1,23 +1,0 @@
-<?php
-
-// List of real names for cisco entities
-$entPhysicalVendorTypes = array(
-    'cevC7xxxIo1feTxIsl'   => 'C7200-IO-FE-MII',
-    'cevChassis7140Dualfe' => 'C7140-2FE',
-    'cevChassis7204'       => 'C7204',
-    'cevChassis7204Vxr'    => 'C7204VXR',
-    'cevChassis7206'       => 'C7206',
-    'cevChassis7206Vxr'    => 'C7206VXR',
-    'cevCpu7200Npe200'     => 'NPE-200',
-    'cevCpu7200Npe225'     => 'NPE-225',
-    'cevCpu7200Npe300'     => 'NPE-300',
-    'cevCpu7200Npe400'     => 'NPE-400',
-    'cevCpu7200Npeg1'      => 'NPE-G1',
-    'cevCpu7200Npeg2'      => 'NPE-G2',
-    'cevPa1feTxIsl'        => 'PA-FE-TX-ISL',
-    'cevPa2feTxI82543'     => 'PA-2FE-TX',
-    'cevPa8e'              => 'PA-8E',
-    'cevPaA8tX21'          => 'PA-8T-X21',
-    'cevMGBIC1000BaseLX'   => '1000BaseLX GBIC',
-    'cevPort10GigBaseLR'   => '10GigBaseLR',
-);
